src/requirements.txt
# OpenAI API for text generation
openai>=1.0.0

# Image processing
Pillow>=10.0.0

# ML/AI frameworks
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# Data processing
numpy>=1.24.0
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0